% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_acc.R
\name{sim_acc}
\alias{sim_acc}
\title{Estimate selection accuracy}
\usage{
sim_acc(pop, SP, geno.file, K.file, n.mate = 200, n.progeny = 50)
}
\arguments{
\item{pop}{parental candidates, AlphaSimR Pop-class}

\item{SP}{simulation parameters for AlphaSimR}

\item{geno.file}{geno filename for COMA}

\item{K.file}{kinship filename for COMA}

\item{n.mate}{number of matings to sample}

\item{n.progeny}{number of progeny per mating}
}
\value{
numeric
}
\description{
Estimate selection accuracy
}
\details{
Details
}
