% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_Gibd.R
\name{sim_Gibd}
\alias{sim_Gibd}
\title{Genomic IBD Matrix in AlphaSimR}
\usage{
sim_Gibd(pop, SP, n.mark, n.core)
}
\arguments{
\item{pop}{AlphaSimR Pop-class}

\item{SP}{simulation parameters for AlphaSimR}

\item{n.mark}{number of markers to sample (evenly spaced in Morgans)}

\item{n.core}{number of cores}
}
\value{
genomic IBD matrix
}
\description{
Genomic IBD Matrix in AlphaSimR
}
\details{
Details
}
