% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_load.R
\name{sim_load}
\alias{sim_load}
\title{Load simulation files}
\usage{
sim_load(rda.file, geno.file, pheno.file, ped.file)
}
\arguments{
\item{rda.file}{R data filename}

\item{geno.file}{genotype CSV file}

\item{pheno.file}{phenotype CSV filename}

\item{ped.file}{pedigree CSV filename}
}
\value{
list containing pop,SP,p.ref
}
\description{
Load simulation files
}
