% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mass.R
\name{sim_mass}
\alias{sim_mass}
\title{Simulate one cycle of mass selection}
\usage{
sim_mass(dF, sel.method, total.progeny, min.progeny, pop, SP, params)
}
\arguments{
\item{dF}{inbreeding rate}

\item{sel.method}{selection method}

\item{total.progeny}{number of progeny for next generation}

\item{min.progeny}{minimum progeny per mating}

\item{pop}{starting population}

\item{SP}{AlphaSimR parameters}

\item{params}{other parameters (Details)}
}
\value{
list containing
\describe{
\item{pop}{Pop-class of progeny}
\item{SP}{simulation parameters}
\item{stats}{list with n.parent, dF, F.A, F.G, accuracy}
}
}
\description{
Simulate one cycle of mass selection
}
\details{
Options for selection method are "PS", "OCS", "OMA", "WGEBV". Results appended to \code{results.file}. Cycle progression: genomic prediction, mate selection, mating, phenotyping, updating TP files. The \code{params} variable is a list with the following: geno.file, pheno.file, K.file, results.file, K.method, gen.TP, gen.record, p.ref, ibd.loci, max.parent, solver, n.core. See \code{\link{sim_TP}}, \code{\link{sim_StageWise}}, \code{\link{sim_OCS}}, \code{\link{sim_OMA}} for details.
}
