% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_OCS.R
\name{sim_OCS}
\alias{sim_OCS}
\title{Parent selection and mating by OCS in AlphaSimR}
\usage{
sim_OCS(
  pop,
  SP,
  n.progeny,
  dF,
  dF.max,
  geno.file,
  K.file,
  min.c = 0.001,
  solver = "ECOS"
)
}
\arguments{
\item{pop}{parental candidates, AlphaSimR Pop-class}

\item{SP}{simulation parameters for AlphaSimR}

\item{n.progeny}{number of progeny to simulate}

\item{dF}{target inbreeding rate}

\item{dF.max}{maximum inbreeding rate}

\item{geno.file}{geno filename for COMA}

\item{K.file}{kinship filename for COMA}

\item{min.c}{minimum contribution}

\item{solver}{CVXR solver}
}
\value{
list containing
\describe{
\item{parents}{Pop-class of selected parents}
\item{ocs}{solution from ocs}
\item{matings}{data frame for sim_mate}
}
}
\description{
Parent selection and mating by OCS in AlphaSimR
}
\details{
Details
}
