% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_OMA.R
\name{sim_OMA}
\alias{sim_OMA}
\title{Parent selection and mating by OMA in AlphaSimR}
\usage{
sim_OMA(
  pop,
  SP,
  n.progeny,
  dF,
  dF.max,
  geno.file,
  K.file,
  max.parent,
  min.a = 0.001,
  solver = "ECOS"
)
}
\arguments{
\item{pop}{parental candidates, AlphaSimR Pop-class}

\item{SP}{simulation parameters for AlphaSimR}

\item{n.progeny}{number of progeny to simulate}

\item{dF}{target inbreeding rate}

\item{dF.max}{maximum inbreeding rate}

\item{geno.file}{geno filename for COMA}

\item{K.file}{kinship filename for COMA}

\item{max.parent}{maximum number of candidate parents for oma}

\item{min.a}{minimum mate allocation}

\item{solver}{CVXR solver}
}
\value{
list containing
\describe{
\item{parents}{Pop-class of selected parents}
\item{oma}{solution from oma}
\item{matings}{data frame for sim_mate}
\item{dev}{mean absolute deviation from random mating}
}
}
\description{
Parent selection and mating by OMA in AlphaSimR
}
\details{
Details
}
