% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_save.R
\name{sim_save}
\alias{sim_save}
\title{Save simulation files}
\usage{
sim_save(rda.file, pop, SP, geno.file, pheno.file, ped.file, p.ref)
}
\arguments{
\item{rda.file}{R data filename}

\item{pop}{variable of AlphaSimR Pop-class}

\item{SP}{simulation parameters for AlphaSimR}

\item{geno.file}{genotype CSV file}

\item{pheno.file}{phenotype CSV filename}

\item{ped.file}{pedigree CSV filename}

\item{p.ref}{reference allele frequencies (for G.VR1)}
}
\description{
Save simulation files
}
