% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StageWise.R
\name{sim_StageWise}
\alias{sim_StageWise}
\title{Estimate marker effects by StageWise}
\usage{
sim_StageWise(
  geno.file,
  pheno.file,
  ploidy,
  COMA.file,
  gen.TP,
  asreml.workspace = "500mb"
)
}
\arguments{
\item{geno.file}{TP genotype CSV filename}

\item{pheno.file}{TP phenotype CSV filename}

\item{ploidy}{ploidy}

\item{COMA.file}{input filename for COMA}

\item{gen.TP}{number of generations for TP (including current)}

\item{asreml.workspace}{memory workspace}
}
\value{
list of dominance parameters
}
\description{
Estimate marker effects by StageWise
}
\details{
Details
}
