% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_TP.R
\name{sim_TP}
\alias{sim_TP}
\title{Add new data to the TP}
\usage{
sim_TP(
  pop,
  SP,
  geno.file,
  pheno.file,
  ped.file,
  K.file,
  K.method,
  gen.record,
  p.ref = NULL,
  ibd.loci = NULL,
  n.core = 1
)
}
\arguments{
\item{pop}{variable of AlphaSimR Pop-class}

\item{SP}{simulation parameters for AlphaSimR}

\item{geno.file}{genotype CSV file}

\item{pheno.file}{phenotype CSV filename}

\item{ped.file}{pedigree CSV filename}

\item{K.file}{kinship CSV filename}

\item{K.method}{kinship method}

\item{gen.record}{number of generations to record (including current)}

\item{p.ref}{reference pop allele frequencies for "VR1" method}

\item{ibd.loci}{number of loci per chromosome for "AM" method}

\item{n.core}{number cores for "AM" method}
}
\value{
list containing
\describe{
\item{F.A}{pedigree inbreeding coefficient}
\item{F.G}{genomic inbreeding coefficient}
\item{p.ref}{allele freq in reference population}
}
}
\description{
Add new data to the TP
}
\details{
Pedigree file has three columns: id, mother, father. Phenotype file has three columns: id, value, gen. To create a new TP, use \code{gen.record} = 0. Argument \code{p.ref} is only used when \code{K.method}="G".
}
