% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trunc.R
\name{sim_trunc}
\alias{sim_trunc}
\title{Truncation selection of parents}
\usage{
sim_trunc(pop, SP, n.progeny, n.parent, geno.file, weighted)
}
\arguments{
\item{pop}{parental candidates, AlphaSimR Pop-class}

\item{SP}{simulation parameters for AlphaSimR}

\item{n.progeny}{number of total progeny}

\item{n.parent}{number of parents}

\item{geno.file}{geno filename with marker effects}

\item{weighted}{TRUE/FALSE for weighting marker effects}
}
\value{
list containing
\describe{
\item{parents}{Pop-class of selected parents}
\item{matings}{data frame for sim_mate}
}
}
\description{
Truncation selection of parents
}
\details{
Details
}
