% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_K.R
\name{update_K}
\alias{update_K}
\title{Update kinship file}
\usage{
update_K(
  pop,
  SP,
  K.method,
  K.file,
  ped.file = NULL,
  p.ref = NULL,
  ibd.loci = 100,
  n.core = 1
)
}
\arguments{
\item{pop}{variable of AlphaSimR Pop-class}

\item{SP}{simulation parameters for AlphaSimR}

\item{K.method}{kinship method: A, G, G.IBD}

\item{K.file}{kinship CSV filename}

\item{ped.file}{pedigree CSV filename}

\item{p.ref}{reference pop allele frequencies for "G" method}

\item{ibd.loci}{number of loci per chromosome for "G.IBD" method}

\item{n.core}{number cores for "G.IBD" method}
}
\value{
average inbreeding coefficient of the population
}
\description{
Update kinship file
}
\details{
Kinship is computed for selection candidates in \code{pop}. For \code{K.method}="A", argument \code{ped.file} is the three-column pedigree file. \code{K.method}="G" specifies VanRaden Method 1, and argument \code{p.ref} is the vector of reference allele frequencies. To reduce computing time for \code{K.method}="G.IBD", only \code{ibd.loci} are sampled per chromosome.
}
