% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_TP.R
\name{update_TP}
\alias{update_TP}
\title{Update the Training Population (TP)}
\usage{
update_TP(pop, SP, geno.file, pheno.file, ped.file, gen.TP)
}
\arguments{
\item{pop}{variable of AlphaSimR Pop-class}

\item{SP}{simulation parameters for AlphaSimR}

\item{geno.file}{genotype CSV file}

\item{pheno.file}{phenotype CSV filename}

\item{ped.file}{pedigree CSV filename}

\item{gen.TP}{# generations recorded for the TP (including current)}
}
\value{
vector of allele frequencies for pop if gen.TP==0, otherwise NULL
}
\description{
Update genotype, genotyped, and pedigree files for the TP
}
\details{
Pedigree file has three columns: id, mother, father. Phenotype file has three columns: id, value, gen. To start a new TP, use \code{gen.TP} = 0.
}
