% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_OCS.R
\name{sim_OCS}
\alias{sim_OCS}
\title{Parent selection and mating by OCS in AlphaSimR}
\usage{
sim_OCS(pop, SP, n.progeny, dF, COMA.file, K.file, solver = "ECOS")
}
\arguments{
\item{pop}{parental candidates, AlphaSimR Pop-class}

\item{SP}{simulation parameters for AlphaSimR}

\item{n.progeny}{number of progeny to simulate}

\item{dF}{target inbreeding rate}

\item{COMA.file}{marker effect or pheno filename for COMA}

\item{K.file}{kinship filename for COMA}

\item{solver}{CVXR solver}
}
\value{
list containing
\describe{
\item{response}{named vector with realized dF, merit, Shannon diversity}
\item{oc}{data frame of optimal contributions for each individual}
\item{om}{data frame of optimal allocations for each mating}
}
}
\description{
Parent selection and mating by OCS in AlphaSimR
}
\details{
Auto detects whether COMA.file has marker effects or phenotype data based on 
the first column name.
}
