% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_OMA.R
\name{sim_OMA}
\alias{sim_OMA}
\title{Parent selection and mating by OMA in AlphaSimR}
\usage{
sim_OMA(
  pop,
  SP,
  n.progeny,
  dF,
  dF.max,
  COMA.file,
  K.file,
  max.parent,
  solver = "ECOS"
)
}
\arguments{
\item{pop}{parental candidates, AlphaSimR Pop-class}

\item{SP}{simulation parameters for AlphaSimR}

\item{n.progeny}{number of progeny to simulate}

\item{dF}{target inbreeding rate}

\item{dF.max}{maximum inbreeding rate}

\item{COMA.file}{geno filename for COMA}

\item{K.file}{kinship filename for COMA}

\item{max.parent}{maximum number of candidate parents for oma}

\item{solver}{CVXR solver}
}
\value{
list containing
\describe{
\item{response}{named vector with realized dF, merit, Shannon diversity}
\item{oc}{data frame of optimal contributions for each individual}
\item{om}{data frame of optimal allocations for each mating}
}
}
\description{
Parent selection and mating by OMA in AlphaSimR
}
\details{
Details
}
