% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StageWise.R
\name{sim_StageWise}
\alias{sim_StageWise}
\title{Estimate marker effects by StageWise within AlphaSimR}
\usage{
sim_StageWise(pop, SP, geno.file)
}
\arguments{
\item{pop}{training population, AlphaSimR Pop-class}

\item{SP}{simulation parameters for AlphaSimR}

\item{geno.file}{geno filename for COMA}
}
\value{
nothing
}
\description{
Estimate marker effects by StageWise within AlphaSimR
}
\details{
Details
}
